Nand basic opt complete(tested). Binary size 1848 bytes(without debug leds). TODO: memory setup...

+++ nand/str.c
@@ -7,5 +7,5 @@ int strcmp(const char *str1, const char *str2)
 		if((res = *str1 - *str2++) != 0 || !*str1++)
 			break;
 	}
-	return (int)res;
+	return res;
 }
